requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
transformers==4.36.2
openai-whisper==20231117
requests==2.31.0
pydantic==2.5.2
python-dotenv==1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
urllib3>=2.0.0
typing-extensions>=4.5.0

